Finish implementing basic xml loader
This commit adds intersection loading, plus adds support for loading of
the start/end of lanes.
Currently, we don't handle (lane) adjacencies properly, but that will come
later.
Appears to be memory-leak free
Does excessively allocate things because of manner in which attributes
are handled (by libxml2). Potential for improvement via different API
call for attribute access
We've replaced *_id types with a templated union; this also has the
retrieve_ptr method that does the lookup into the reference maps
Finally, explicitly add ltstr struct for map comparison function
Guess it was using pointers and not some automatic semantics, whoops.